src/bookmark/file_manager.py
"""File management utilities for the bookmark manager."""

import sys
from pathlib import Path

import yaml


class BookmarkFileManager:
    """Manages bookmark directory structure and file operations.

    Handles the ~/.bookmarks directory structure, ensuring proper setup
    and providing utilities for file operations within the project constraints.
    """

    def __init__(self, custom_base_dir: Path | None = None) -> None:
        """Initialize the file manager.

        Args:
            custom_base_dir: Custom base directory for testing (defaults to ~/.bookmarks)
        """
        if custom_base_dir:
            self.base_dir = custom_base_dir
        else:
            # For actual use, this would be ~/.bookmarks, but for testing we simulate it
            # within the project directory to respect the ABSOLUTE IMPERATIVES
            home = Path.home()
            self.base_dir = home / ".bookmarks"

        self.tags_file = self.base_dir / "tags.txt"
        self.config_file = self.base_dir / "config.yaml"

    def ensure_base_directory(self) -> None:
        """Ensure the base bookmarks directory exists.

        Raises:
            PermissionError: If directory cannot be created
        """
        try:
            self.base_dir.mkdir(parents=True, exist_ok=True)
        except PermissionError as e:
            print(
                f"Error: Cannot create bookmarks directory {self.base_dir}: {e}",
                file=sys.stderr,
            )
            sys.exit(1)
        except OSError as e:
            print(
                f"Error: Failed to create bookmarks directory {self.base_dir}: {e}",
                file=sys.stderr,
            )
            sys.exit(1)

    def get_bookmark_file_path(self, filename: str | None = None) -> Path:
        """Get the path to a bookmark file.

        Args:
            filename: Optional filename, defaults to "bookmarks.txt"

        Returns:
            Path to the bookmark file
        """
        if filename is None:
            filename = "bookmarks.txt"

        # If filename is just a name (no path), put it in the base directory
        if "/" not in filename and "\\" not in filename:
            return self.base_dir / filename
        # If it's a path, treat it as relative to base directory
        return self.base_dir / filename

    def read_tags(self) -> list[str]:
        """Read available tags from tags.txt.

        Returns:
            List of available tags sorted alphabetically
        """
        if not self.tags_file.exists():
            return []

        try:
            with open(self.tags_file, encoding="utf-8") as f:
                tags = [line.strip().lower() for line in f if line.strip()]
                return sorted(set(tags))  # Remove duplicates and sort
        except (PermissionError, OSError) as e:
            print(f"Warning: Cannot read tags file: {e}", file=sys.stderr)
            return []

    def update_tags(self, new_tags: list[str]) -> None:
        """Update the tags file with new tags.

        Args:
            new_tags: List of new tags to add
        """
        if not new_tags:
            return

        # Read existing tags
        existing_tags = self.read_tags()

        # Add new tags (normalized to lowercase)
        all_tags = existing_tags + [tag.lower().strip() for tag in new_tags]
        unique_tags = sorted(set(all_tags))

        # Write back to file
        try:
            self.ensure_base_directory()
            with open(self.tags_file, "w", encoding="utf-8") as f:
                f.writelines(f"{tag}\n" for tag in unique_tags)
        except (PermissionError, OSError) as e:
            print(f"Warning: Cannot update tags file: {e}", file=sys.stderr)

    def read_config(self) -> dict:
        """Read configuration from config.yaml.

        Returns:
            Configuration dictionary with defaults
        """
        default_config = {
            "display_fields": ["name", "description", "url"],
            "browser": None,
            "default_bookmark_file": "bookmarks.txt",
        }

        if not self.config_file.exists():
            return default_config

        try:
            with open(self.config_file, encoding="utf-8") as f:
                config = yaml.safe_load(f)
                if config is None:
                    return default_config

                # Merge with defaults
                result = default_config.copy()
                result.update(config)
                return result
        except (PermissionError, OSError) as e:
            print(f"Warning: Cannot read config file: {e}", file=sys.stderr)
            return default_config
        except yaml.YAMLError as e:
            print(f"Warning: Invalid YAML in config file: {e}", file=sys.stderr)
            return default_config


class ProjectTestFileManager(BookmarkFileManager):
    """Test-specific file manager that operates within project directory.

    This manager respects the ABSOLUTE IMPERATIVE to never modify files
    outside the project directory by creating a mock ~/.bookmarks structure
    within the project's test fixtures.
    """

    def __init__(self, test_base_dir: Path) -> None:
        """Initialize with a test base directory within the project.

        Args:
            test_base_dir: Base directory for test bookmarks (within project)
        """
        super().__init__(custom_base_dir=test_base_dir)
